+++ test/csrf_cookieDomain.test.js
@@ -44,4 +44,44 @@ describe('test/csrf_cookieDomain.test.js', () => {
         .expect('Set-Cookie', /csrfToken=[\w\-]+; path=\/; domain=\.string\.com/);
     });
   });
+
+  describe('cookieOptions = object', () => {
+    let app;
+    before(() => {
+      app = mm.app({
+        baseDir: 'apps/csrf-cookieOptions',
+      });
+      return app.ready();
+    });
+    after(() => app.close());
+
+    it('should auto set csrfToken with cookie options on GET request', () => {
+      return app.httpRequest()
+        .get('/hello')
+        .set('Host', 'abc.aaaa.ddd.string.com')
+        .expect('hello csrfToken cookieOptions')
+        .expect(200)
+        .expect('Set-Cookie', /csrfToken=[\w\-]+; path=\/; httponly/);
+    });
+  });
+
+  describe('cookieOptions use signed', () => {
+    let app;
+    before(() => {
+      app = mm.app({
+        baseDir: 'apps/csrf-cookieOptions-signed',
+      });
+      return app.ready();
+    });
+    after(() => app.close());
+
+    it('should auto set csrfToken and csrfToken.sig with cookie options on GET request', () => {
+      return app.httpRequest()
+        .get('/hello')
+        .set('Host', 'abc.aaaa.ddd.string.com')
+        .expect('hello csrfToken cookieOptions signed')
+        .expect(200)
+        .expect('Set-Cookie', /csrfToken=[\w\-]+; path=\/,csrfToken\.sig=[\w\-]+; path=\//);
+    });
+  });
 });